+++ docs/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - feat: allow for setting HttpAgent ingress expiry using `ingressExpiryInMinutes` option
 
+- feat: improved assertion options for agent errors using `prototype`, `name`, and `instanceof`
+
 ### Changed
 
 - test: automatically deploys trap canister if it doesn't exist yet during e2e